refactor: move dashboards widgets
