+++ lib/beamer/locations/dashboards_location.dart
@@ -1,7 +1,7 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
-import 'package:lotti/pages/dashboards/dashboard_page.dart';
-import 'package:lotti/pages/dashboards/dashboards_list_page.dart';
+import 'package:lotti/features/dashboards/ui/pages/dashboard_page.dart';
+import 'package:lotti/features/dashboards/ui/pages/dashboards_list_page.dart';
 import 'package:lotti/utils/uuid.dart';
 
 class DashboardsLocation extends BeamLocation<BeamState> {
