+++ lib/features/dashboards/ui/widgets/dashboards_list.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/dashboards/dashboards_page_cubit.dart';
 import 'package:lotti/blocs/dashboards/dashboards_page_state.dart';
-import 'package:lotti/widgets/dashboards/dashboards_card.dart';
+import 'package:lotti/features/dashboards/ui/widgets/dashboards_card.dart';
 
 class DashboardsList extends StatelessWidget {
   const DashboardsList({
